move modal to shared location
