+++ src/app/core/core.component.html
@@ -23,3 +23,5 @@
 <app-subscribe-toast *ngIf="showSubscribeToast" (emitCloseToast)="hideSubscribeToast()"></app-subscribe-toast>
 
 <app-snackbar-notification></app-snackbar-notification>
+
+<app-export-to-justifi-modal *ngIf="showExportToJustifiModal"></app-export-to-justifi-modal>
