+++ src/app/core/core.component.ts
@@ -20,12 +20,13 @@ import { ImportBackupModalService } from '../shared/import-backup-modal/import-b
 import { MeasurSurveyService } from '../shared/measur-survey/measur-survey.service';
 import { UpdateApplicationService } from '../shared/update-application/update-application.service';
 import { EmailListSubscribeService } from '../shared/subscribe-toast/email-list-subscribe.service';
+import { ExportToJustifiTemplateService } from '../shared/export-to-justifi-modal/export-to-justifi-services/export-to-justifi-template.service';
 
 @Component({
-    selector: 'app-core',
-    templateUrl: './core.component.html',
-    styleUrls: ['./core.component.css'],
-    standalone: false
+  selector: 'app-core',
+  templateUrl: './core.component.html',
+  styleUrls: ['./core.component.css'],
+  standalone: false
 })
 
 export class CoreComponent implements OnInit {
@@ -57,6 +58,8 @@ export class CoreComponent implements OnInit {
   showSubscribeToastSub: Subscription;
   subscribeModalSub: Subscription;
   emailVisibilitySubscription: Subscription;
+  showExportToJustifiModal: boolean = false;
+  showExportToJustifiModalSub: Subscription;
 
   constructor(public electronService: ElectronService,
     private assessmentService: AssessmentService,
@@ -78,7 +81,8 @@ export class CoreComponent implements OnInit {
     private measurSurveyService: MeasurSurveyService,
     private updateApplicationService: UpdateApplicationService,
     private emailSubscribeService: EmailListSubscribeService,
-    private inventoryDbService: InventoryDbService) {
+    private inventoryDbService: InventoryDbService,
+    private exportToJustifiTemplateService: ExportToJustifiTemplateService) {
   }
 
   ngOnInit() {
@@ -155,6 +159,10 @@ export class CoreComponent implements OnInit {
     this.showImportBackupModalSubscription = this.importBackupModalService.showImportBackupModal.subscribe(showModal => {
       this.showImportBackupModal = showModal;
     });
+
+    this.showExportToJustifiModalSub = this.exportToJustifiTemplateService.showExportToJustifiModal.subscribe(val => {
+      this.showExportToJustifiModal = val;
+    });
   }
 
 
@@ -203,11 +211,11 @@ export class CoreComponent implements OnInit {
         await firstValueFrom(this.applicationInstanceDbService.setSurveyDone());
       } else {
         let hasMetModalRequirements = this.measurSurveyService.getHasModalUsageRequirements(applicationData);
-        
+
         setTimeout(() => {
           this.measurSurveyService.showSurveyModal.next(hasMetModalRequirements);
         }, 5000);
-        
+
         let canShowToast = this.measurSurveyService.getHasToastUsageRequirements(applicationData);
         if (canShowToast && !applicationData.isSurveyToastDone && !hasMetModalRequirements) {
           setTimeout(() => {
@@ -224,7 +232,7 @@ export class CoreComponent implements OnInit {
   }
 
   setAllDbData() {
-      this.coreService.getAllAppData()
+    this.coreService.getAllAppData()
       .pipe(catchError(error => this.appErrorService.handleObservableAppError('Error loading MEASUR database', error))).subscribe({
         next: (initializedData) => {
           this.directoryDbService.setAll(initializedData.directories);
@@ -240,13 +248,13 @@ export class CoreComponent implements OnInit {
             this.automaticBackupService.saveVersionedBackup();
           }
         },
-        error: (error) => {}
+        error: (error) => { }
       });
   }
 
   async hideSurveyToast() {
     this.setSurveyToastDone();
-  }  
+  }
 
   async setSurveyToastDone() {
     this.showSurveyToast = false;
