+++ src/app/dashboard/directory-dashboard/directory-dashboard.component.html
@@ -43,4 +43,3 @@
 
 <app-delete-items-modal *ngIf="showDeleteItemsModal"></app-delete-items-modal>
 <app-pre-assessment-modal *ngIf="showPreAssessmentModalIndex != undefined"></app-pre-assessment-modal>
-<app-export-to-justifi-modal *ngIf="showExportToJustifiModal"></app-export-to-justifi-modal>
