+++ src/app/dashboard/directory-dashboard/directory-dashboard.component.ts
@@ -36,8 +36,6 @@ export class DirectoryDashboardComponent implements OnInit {
   sortBySub: Subscription;
 
   dashboardCollapsed: boolean = false;
-  showExportToJustifiModal: boolean = false;
-  showExportToJustifiModalSub: Subscription;
   constructor(private activatedRoute: ActivatedRoute, private directoryDbService: DirectoryDbService,
     private directoryDashboardService: DirectoryDashboardService, private dashboardService: DashboardService,
     private settingsDbService: SettingsDbService, private assessmentService: AssessmentService) { }
@@ -79,9 +77,6 @@ export class DirectoryDashboardComponent implements OnInit {
     this.sortBySub = this.directoryDashboardService.sortBy.subscribe(val => {
       this.sortBy = val;
     });
-    this.showExportToJustifiModalSub = this.directoryDashboardService.showExportToJustifiModal.subscribe(val => {
-      this.showExportToJustifiModal = val;
-    });
 
     if (!this.settingsDbService.globalSettings.disableDashboardTutorial) {
       this.assessmentService.showTutorial.next('dashboard-tutorial');
@@ -95,7 +90,6 @@ export class DirectoryDashboardComponent implements OnInit {
     this.updateDashboardDataSub.unsubscribe();
     this.filterDashboardBySub.unsubscribe();
     this.sortBySub.unsubscribe();
-    this.showExportToJustifiModalSub.unsubscribe();
   }
 
   setDirectory() {
