+++ src/app/shared/export-to-justifi-modal/export-to-justifi-modal.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ExportToJustifiModalComponent } from './export-to-justifi-modal.component';
+import { ModalModule } from 'ngx-bootstrap/modal';
+
+
+
+@NgModule({
+  declarations: [
+    ExportToJustifiModalComponent
+  ],
+  imports: [
+    CommonModule,
+    ModalModule
+  ],
+  exports: [
+    ExportToJustifiModalComponent]
+})
+export class ExportToJustifiModalModule { }
