+++ src/app/treasure-hunt/treasure-hunt-report/treasure-hunt-report.component.css
@@ -13,8 +13,13 @@
 .date-modified {
     color: #999;
     font-size: 12px;
-    margin-top: 15px;
-    margin-left: 0;
+    /* margin-top: 15px;
+    margin-left: 0; */
+}
+
+.mt-auto{
+    margin-top: auto !important;
+    margin-bottom: 0;
 }
 
 .report-diagrams {
